+++ app/components/primer/primer.pcss
@@ -33,5 +33,5 @@
 @import "./beta/truncate.pcss";
 
 /* to be renamed */
-@import "./timeline_item_component.pcss";
+@import "./beta/timeline_item.pcss";
 @import "./truncate.pcss";
