+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -104,7 +104,7 @@
   - title: TextField
     url: "/components/alpha/textfield"
   - title: TimelineItem
-    url: "/components/timelineitem"
+    url: "/components/beta/timelineitem"
   - title: ToggleSwitch
     url: "/components/alpha/toggleswitch"
   - title: Tooltip
