+++ static/statuses.json
@@ -67,6 +67,8 @@
   "Primer::Beta::State": "beta",
   "Primer::Beta::Subhead": "beta",
   "Primer::Beta::Text": "beta",
+  "Primer::Beta::TimelineItem": "beta",
+  "Primer::Beta::TimelineItem::BadgeComponent": "alpha",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
   "Primer::BlankslateComponent": "deprecated",
@@ -97,8 +99,7 @@
   "Primer::SubheadComponent": "deprecated",
   "Primer::TabContainerComponent": "deprecated",
   "Primer::TimeAgoComponent": "deprecated",
-  "Primer::TimelineItemComponent": "beta",
-  "Primer::TimelineItemComponent::BadgeComponent": "alpha",
+  "Primer::TimelineItemComponent": "deprecated",
   "Primer::Tooltip": "deprecated",
   "Primer::Truncate": "beta"
 }
